Add consumer service Up/uOwn name in metrics
Summary:
This diff adds the owning service name into throughput and quota metrics.
This Up/uOwn service name is for UAE to identify the service for sending alerts to.
It is not related to the RPC service name, which is solely used by the Go
consumer services to identify the requests to its GoFx gRPC server.