[RpcColocatingRebalancer] Rebuild worker -> partition mapping from ZK (Stacked Diff 1/2)
Summary:
Previously, after finding the # of worker count needed per virtual partition, we do one of the following:

1.) If rebalancing cache is stale, add workers "randomly" to each partition according to its needs
2.) If not stale, we get rid of stale workers in the partition and add some extra workers

This means that if you restart controller, you lose the in memory cache and the workers will get shuffled

This diff tries to combine 1 and 2, since partition is based on job group, we try to get the workers in a partition via the worker ID in the job group. We get rid of any stale ones. I will rename rebalancing cache -> rebalancing table in a seperate diff to make this diff easier to review, it's not really a cache anymore.

Test Plan:
Unit test (needs to be adjusted to have a higher worker to reduce ratio/more rebalancer runs to converge, because now on the controller restart, the workers get shrinked in batches of 10% by default)

 dev testing
1.) Deploy colocating rebalancer prior to this build, then deploy this diff, jobs don't get shuffled
2.) Restart this rebalancer, jobs don't get shuffled
3.) Deploy streamingRpcUri rebalancer, then this one, rebalancer doesn't keep shuffling jobs
4.) add some job, worker count went up/remove job worker count goes down