[Kafka/uForwarder]Job comparison should use jobId instead of Job object
Summary:
[Kafka/uForwarder]Job comparison should use jobId instead of Job object

Today in the AbstractKafkaFetcherThread, we compare the new jobs and old jobs to determine whether the assignment is changed and make some logging. However, the current logic is using the Job object in the Set to do the comparison which is not correct because if a job is updated for the quota, it will still be considered as assignment change.

The diff is to fix this behavior.