retrieve commit offset on fetcher thread startup
Summary:

A previous diff allows the fetcher thread to commit the offset even if the offset is not moving forward. One corner case is: commit offset will not be cached in memory for offset commit.
This diff will fix the corner case by fetching the commit offset stored in the broker on fetcher startup.
 A new feature flag is added for gradual rollout; it will be removed once the feature is fully rolled out.

in dev environment: worker starts to commit offset after the change get deployed,  lag report is normal

dev-dlq environment: worker starts to commit offset on dlq merge and the offset commit stopped upon merge completion