RpcJobColocatingRebalancer should move extra overloaded jobs to idle workers to balance the load
Summary:
RpcJobColocatingRebalancer should move extra overloaded jobs to idle workers to balance the load

Currently if there is a worker holding more than 1 jobs with workload > 1.0, and if there are idle workers, these jobs with workload>1.0 will not be moved to the idle workers.

This diff is adding another check so that the extra overloaded job will be moved to idle workers.

Test Plan: Unit test with production data.