Should pack workers when esitimating worker number
Summary:
Should pack workers when esitimating worker number

Today when estimating the needed worker number for a virtual partition, we are using a greedy algorithm which can't pack the small jobs with the previously-visited jobs and cause more workers used than needed.

We are changing it to use Heap-based solution so that we can also pack the jobs when estimating the worker number. Meanwhile, when rebalancing the stale jobs(whose workers are dead), we are also applying the Heap-based solution to place the heavy job first.