[kafka/KCP] Create root span for messages without trace IDs
Summary:
Currently, SWN clients will create root span if there's no trace IDs.
However, if KCP don't create root span when the message is missing trace
ID, it's impossible for the SWN client to distinguish which part actually
went wrong. To reduce the effort on the consumer service side, we decided
to create root span so we narrow the problem to the producer.