[uforwarder] fix build and test failure in uforwarder microrepo
Summary:
(1)
With the current protoc and protoc-gen-grpc-java version, it cannot compile on Mac computers with Apple silicon (M1, M2, M3).
Compile Errors

Execution failed for task ':idl:generateProto'.
> Could not resolve all files for configuration ':idl:protobufToolsLocator_protoc'.
   > Could not find protoc-3.6.1-osx-aarch_64.exe (com.google.protobuf:protoc:3.6.1).
Execution failed for task ':idl:generateProto'.
> Could not resolve all files for configuration ':idl:protobufToolsLocator_grpc'.
   > Could not find protoc-gen-grpc-java-1.15.1-osx-aarch_64.exe (io.grpc:protoc-gen-grpc-java:1.15.1).
Both the above libraries cannot be found under

https://repo1.maven.org/maven2/com/google/protobuf/protoc/3.6.1/
https://repo1.maven.org/maven2/io/grpc/protoc-gen-grpc-java/1.15.1/
but can found in

https://repo1.maven.org/maven2/com/google/protobuf/protoc/3.18.0/
https://repo1.maven.org/maven2/io/grpc/protoc-gen-grpc-java/1.50.0/

(2)
Remove DelayProcessingManagerNoopTest, as with it, the KafkaCheckpointManagerTest fail with mockito error, it requires to update mockito to 5.0.x. As java monorepo right now, support 4.5.x. We delete the NoopTest as it didn't as much value.

Reviewers: #ldap_owner_kafka_monorepo_blocking, #ldap_owner_kafka_monorepo_nonblocking, yayang, chenz

Reviewed By: #ldap_owner_kafka_monorepo_blocking, #ldap_owner_kafka_monorepo_nonblocking, chenz

JIRA Issues: KAFEP-7231

Differential Revision: https://code.uberinternal.com/D15867365