[kafka/KCP] Add ScaleStatus to AddJobGroupRequest
Summary:
We are going to revert the change in D17310079 because it's very
weird to have 2 ScaleStatus in the StoredJobGroup. We combine the
code and IDL in the single diff so we know how it works, and it
also guarantees that the IDL change won't break.

Test Plan:
UT

Add Request in Studio: https://api-explorer.uberinternal.com/b1ea5804-8e3a-44fb-847b-4979acdeccb8

Get response after add
{F1072387239}

Reviewers: #ldap_owner_kafka_monorepo_blocking, #ldap_owner_kafka_monorepo_nonblocking, ureview, hf

Reviewed By: #ldap_owner_kafka_monorepo_blocking, #ldap_owner_kafka_monorepo_nonblocking, hf

Tags: #backward_incompatible_idl

JIRA Issues: KAFEP-7768

Differential Revision: https://code.uberinternal.com/D17311729